Fix for CHEF-1571, reload cookbook metadata after generating metadata
CookbookVersion objects now have the following new accessors:
  root_dir
  metadata_json_file
  metadata_rb_file

The CookbookLoader sets a CookbookVersion's root_dir on load.

CookbookVersion objects now have a reload_metadata! method that reloads
metadata from metadata.json (if it exists).  This is used by the
CookbookUploader to reload a cookbook's metadata after (re)generating
metadata in prep for upload.

Thanks to Dan Prince for the details reported in CHEF-1560 and for a
patch that inspired these changes.